Edited version of selscan-main
Added a new function calculateHomozygosity_Wagh, which, in addition to haplotype counts and total number of chromosomes, takes in the number of chromosomes with allele '1' at the core SNP. EHH is then calculated with the denominator nCk(N_1, 2)+nCk(N_2, 2), where N_1 and N_2 are the numbers of chromosomes in the sample with each allele at the core SNP. This is following Wagh et al. 2012 (PLoS One). The initialization step of XP-EHH also reflects this change.